Reduce warnings Untested implementation of sprint Partly working physics Allow smaller maps of offsets in both dimensions Passthrough program name via build system

+++ widgets.c
@@ -10,7 +10,7 @@ char menu(char **items, char amount, char start_line, char offset, MenuMode mode
         char line = start_line; 
         
         for (char n=0; n < amount; ++n) {
-            if (!((previous == -1) || (n == option) || (n == previous)))
+            if (!((previous == -1) || (n == option) || ((signed int)n == (signed int)previous)))
                 continue;    
             
             strcpy(output, "  ");
